REF: combine IndexEngine test files
